refactor: Split the main agent.go file into logical units for clarity improvements
Still thinking how I should generate those files, it's a design choice that helps creating more agents.
